+++ packages/core/src/components.d.ts
@@ -21,6 +21,7 @@ import { ExpandableAnimationEndEvent, ExpandableAnimationStartEvent } from "./co
 import { HeaderEvent, HeaderMenuItem } from "./components/header/header.interfaces";
 import { InfoButtonToggleEvent } from "./components/info-button/info-button.interfaces";
 import { InputRangeChangeEvent } from "./components/input-range/input-range.interfaces";
+import { LegendItemRemoveClickEvent } from "./components/legend-item/legend-item.interfaces";
 import { ListButtonChangeEvent, ListButtonSelectedEvent } from "./components/list-button/list-button.interfaces";
 import { BaseLayer, BaseLayerChangeEvent } from "./components/map-base-layers/map-base-layers.interfaces";
 import { MapControlsToggleEvent } from "./components/map-controls/map-controls.interfaces";
@@ -51,6 +52,7 @@ export { ExpandableAnimationEndEvent, ExpandableAnimationStartEvent } from "./co
 export { HeaderEvent, HeaderMenuItem } from "./components/header/header.interfaces";
 export { InfoButtonToggleEvent } from "./components/info-button/info-button.interfaces";
 export { InputRangeChangeEvent } from "./components/input-range/input-range.interfaces";
+export { LegendItemRemoveClickEvent } from "./components/legend-item/legend-item.interfaces";
 export { ListButtonChangeEvent, ListButtonSelectedEvent } from "./components/list-button/list-button.interfaces";
 export { BaseLayer, BaseLayerChangeEvent } from "./components/map-base-layers/map-base-layers.interfaces";
 export { MapControlsToggleEvent } from "./components/map-controls/map-controls.interfaces";
@@ -665,6 +667,20 @@ export namespace Components {
          */
         "truncate"?: boolean;
     }
+    interface DsoLegendItem {
+        /**
+          * To disable the Legend Item
+         */
+        "disabled": boolean;
+        /**
+          * Message to be shown behind a toggletip when the Legend Item is disabled
+         */
+        "disabledMessage"?: string;
+        /**
+          * Shows a trash-can that, when clicked, emits `dsoRemoveClick`.
+         */
+        "removable"?: boolean;
+    }
     interface DsoListButton {
         /**
           * Whether the List Button is checked.
@@ -1126,6 +1142,10 @@ export interface DsoLabelCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLDsoLabelElement;
 }
+export interface DsoLegendItemCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLDsoLegendItemElement;
+}
 export interface DsoListButtonCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLDsoListButtonElement;
@@ -1538,6 +1558,25 @@ declare global {
         prototype: HTMLDsoLabelElement;
         new (): HTMLDsoLabelElement;
     };
+    interface HTMLDsoLegendItemElementEventMap {
+        "dsoRemoveClick": LegendItemRemoveClickEvent;
+        "dsoMouseEnter": any;
+        "dsoMouseLeave": any;
+    }
+    interface HTMLDsoLegendItemElement extends Components.DsoLegendItem, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLDsoLegendItemElementEventMap>(type: K, listener: (this: HTMLDsoLegendItemElement, ev: DsoLegendItemCustomEvent<HTMLDsoLegendItemElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLDsoLegendItemElementEventMap>(type: K, listener: (this: HTMLDsoLegendItemElement, ev: DsoLegendItemCustomEvent<HTMLDsoLegendItemElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLDsoLegendItemElement: {
+        prototype: HTMLDsoLegendItemElement;
+        new (): HTMLDsoLegendItemElement;
+    };
     interface HTMLDsoListButtonElementEventMap {
         "dsoCountChange": ListButtonChangeEvent;
         "dsoSelectedChange": ListButtonSelectedEvent;
@@ -1876,6 +1915,7 @@ declare global {
         "dso-info-button": HTMLDsoInfoButtonElement;
         "dso-input-range": HTMLDsoInputRangeElement;
         "dso-label": HTMLDsoLabelElement;
+        "dso-legend-item": HTMLDsoLegendItemElement;
         "dso-list-button": HTMLDsoListButtonElement;
         "dso-logo": HTMLDsoLogoElement;
         "dso-map-base-layers": HTMLDsoMapBaseLayersElement;
@@ -2609,6 +2649,32 @@ declare namespace LocalJSX {
          */
         "truncate"?: boolean;
     }
+    interface DsoLegendItem {
+        /**
+          * To disable the Legend Item
+         */
+        "disabled"?: boolean;
+        /**
+          * Message to be shown behind a toggletip when the Legend Item is disabled
+         */
+        "disabledMessage"?: string;
+        /**
+          * Emitted when the mouse enters the Legend Item
+         */
+        "onDsoMouseEnter"?: (event: DsoLegendItemCustomEvent<any>) => void;
+        /**
+          * Emitted when the mouse leaves the Legend Item
+         */
+        "onDsoMouseLeave"?: (event: DsoLegendItemCustomEvent<any>) => void;
+        /**
+          * Emitted when the user activates the remove button.
+         */
+        "onDsoRemoveClick"?: (event: DsoLegendItemCustomEvent<LegendItemRemoveClickEvent>) => void;
+        /**
+          * Shows a trash-can that, when clicked, emits `dsoRemoveClick`.
+         */
+        "removable"?: boolean;
+    }
     interface DsoListButton {
         /**
           * Whether the List Button is checked.
@@ -3136,6 +3202,7 @@ declare namespace LocalJSX {
         "dso-info-button": DsoInfoButton;
         "dso-input-range": DsoInputRange;
         "dso-label": DsoLabel;
+        "dso-legend-item": DsoLegendItem;
         "dso-list-button": DsoListButton;
         "dso-logo": DsoLogo;
         "dso-map-base-layers": DsoMapBaseLayers;
@@ -3191,6 +3258,7 @@ declare module "@stencil/core" {
             "dso-info-button": LocalJSX.DsoInfoButton & JSXBase.HTMLAttributes<HTMLDsoInfoButtonElement>;
             "dso-input-range": LocalJSX.DsoInputRange & JSXBase.HTMLAttributes<HTMLDsoInputRangeElement>;
             "dso-label": LocalJSX.DsoLabel & JSXBase.HTMLAttributes<HTMLDsoLabelElement>;
+            "dso-legend-item": LocalJSX.DsoLegendItem & JSXBase.HTMLAttributes<HTMLDsoLegendItemElement>;
             "dso-list-button": LocalJSX.DsoListButton & JSXBase.HTMLAttributes<HTMLDsoListButtonElement>;
             "dso-logo": LocalJSX.DsoLogo & JSXBase.HTMLAttributes<HTMLDsoLogoElement>;
             "dso-map-base-layers": LocalJSX.DsoMapBaseLayers & JSXBase.HTMLAttributes<HTMLDsoMapBaseLayersElement>;