+++ packages/dso-toolkit/src/components/legend-item/legend-item.models.ts
@@ -1,0 +1,15 @@
+export interface LegendItem<TemplateFnReturnType> {
+  body?: TemplateFnReturnType;
+  content: TemplateFnReturnType | string;
+  disabled?: boolean;
+  disabledMessage?: string;
+  dsoMouseEnter?: () => void;
+  dsoMouseLeave?: () => void;
+  dsoRemoveClick?: (e: CustomEvent<LegendItemRemoveClickEvent>) => void;
+  removable?: boolean;
+  symbol?: TemplateFnReturnType;
+}
+
+export interface LegendItemRemoveClickEvent {
+  originalEvent: MouseEvent;
+}