+++ storybook/src/components/legend-item/legend-item.content.ts
@@ -1,0 +1,3 @@
+import { html } from "lit-html";
+
+export const defaultSymbol = html`<span class="symboolcode" data-symboolcode="regelingsgebied"></span>`;